make genesis tools standalone
